requirements.txt
streamlit==1.43.0
groq
python-dotenv==1.0.1
